Added a new feature
Add HTML structure for displaying a timeline of activities in a container with a class name “container”. The timeline consists of multiple items with a class name “box-item3”. Each item has a body with a class name “box-body”.

Inside the body, there is a text box with a class name “text-box3”. It contains a title with a class name “t-b-3” and an image of a list.

Next to the text box, there is an icon box with a class name “icone-box”. It contains an SVG icon of three vertical dots.

Underneath the text box, there is a main text section with a class name “main-text”. It consists of multiple sections with different class names “main-text1”, “main-text2”, “main-text3”.

Each section has a title and date. Inside the first section “main-text1”, there is a paragraph describing a meeting with a customer. Below that, there is an image and a paragraph representing the customer’s information.

In the second section “main-text2”, there is a paragraph describing the creation of a new project for the customer.

In the third section “main-text3”, there is a paragraph and an image representing a message sent by Tony Stark. There are also two more images representing a PDF file and a document, along with their respective captions.

In the last section “main-text3”, there is a paragraph describing some content.

This is the GitHub commit message describing the addition of the HTML structure for displaying the timeline of activities in the container.